chore: partially componentized
